- changed main activity to a splash screen - added level1 link to the recyclerview quiz - start quiz links back to the recycler view
