+++ app/src/main/res/layout/activity_quiz.xml
@@ -6,16 +6,9 @@
     android:layout_height="match_parent"
     tools:context=".quiz">
 
-    <TextView
-        android:id="@+id/startQuiz"
-        android:layout_width="173dp"
-        android:layout_height="109dp"
-        android:layout_marginTop="260dp"
-        android:layout_marginEnd="136dp"
-        android:layout_marginRight="136dp"
-        android:text="quiz here"
-        android:textAppearance="@style/TextAppearance.AppCompat.Body2"
-        android:textSize="30sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:scrollbars="vertical" />
 </androidx.constraintlayout.widget.ConstraintLayout>
